requirements.txt
streamlit
music21
python-dotenv
langchain
langchain_core
langchain_community
langchain_groq
scipy
synthesizer
